+++ spec/psn/client/trophies_spec.rb
@@ -64,6 +64,38 @@ RSpec.describe PSN::Client::Trophies do
       expect(request).to have_received(:get).with(path)
     end
 
+    it 'derives service name for PS3 platform' do
+      path = "/trophy/v1/npCommunicationIds/#{np_comm_id}/trophyGroups/all/trophies?npServiceName=trophy"
+
+      trophies.title_trophies(np_communication_id: np_comm_id, platform: 'PS3')
+
+      expect(request).to have_received(:get).with(path)
+    end
+
+    it 'derives service name for PSVita platform' do
+      path = "/trophy/v1/npCommunicationIds/#{np_comm_id}/trophyGroups/all/trophies?npServiceName=trophy"
+
+      trophies.title_trophies(np_communication_id: np_comm_id, platform: 'PSVita')
+
+      expect(request).to have_received(:get).with(path)
+    end
+
+    it 'derives service name for PC platform' do
+      path = "/trophy/v1/npCommunicationIds/#{np_comm_id}/trophyGroups/all/trophies?npServiceName=trophy2"
+
+      trophies.title_trophies(np_communication_id: np_comm_id, platform: 'PC')
+
+      expect(request).to have_received(:get).with(path)
+    end
+
+    it 'accepts a custom trophy_group_id' do
+      path = "/trophy/v1/npCommunicationIds/#{np_comm_id}/trophyGroups/001/trophies?"
+
+      trophies.title_trophies(np_communication_id: np_comm_id, trophy_group_id: '001')
+
+      expect(request).to have_received(:get).with(path)
+    end
+
     it 'raises an error if both np_service_name and platform are provided' do
       expect do
         trophies.title_trophies(np_communication_id: np_comm_id, platform: 'PS4', np_service_name: 'trophy')
@@ -74,6 +106,93 @@ RSpec.describe PSN::Client::Trophies do
       path = "/trophy/v1/npCommunicationIds/#{np_comm_id}/trophyGroups/all/trophies?limit=10&offset=5"
 
       trophies.title_trophies(np_communication_id: np_comm_id, limit: 10, offset: 5)
+
+      expect(request).to have_received(:get).with(path)
+    end
+  end
+
+  describe '#earned_trophies' do
+    let(:response) { { 'trophies' => [] } }
+    let(:np_comm_id) { 'NPWR12345_00' }
+
+    before do
+      allow(PSN::Client::Request).to receive(:new).with(access_token).and_return(request)
+      allow(request).to receive(:get).and_return(response)
+    end
+
+    it 'calls the correct endpoint with default values' do
+      path = "/trophy/v1/users/me/npCommunicationIds/#{np_comm_id}/trophyGroups/all/trophies?"
+
+      expect(trophies.earned_trophies(np_communication_id: np_comm_id)).to eq(response)
+      expect(request).to have_received(:get).with(path)
+    end
+
+    it 'accepts a custom user_id' do
+      path = "/trophy/v1/users/12345/npCommunicationIds/#{np_comm_id}/trophyGroups/all/trophies?"
+
+      trophies.earned_trophies(user_id: '12345', np_communication_id: np_comm_id)
+
+      expect(request).to have_received(:get).with(path)
+    end
+
+    it 'derives service name for PS4 platform' do
+      path = "/trophy/v1/users/me/npCommunicationIds/#{np_comm_id}/trophyGroups/all/trophies?npServiceName=trophy"
+
+      trophies.earned_trophies(np_communication_id: np_comm_id, platform: 'PS4')
+
+      expect(request).to have_received(:get).with(path)
+    end
+
+    it 'derives service name for PS5 platform' do
+      path = "/trophy/v1/users/me/npCommunicationIds/#{np_comm_id}/trophyGroups/all/trophies?npServiceName=trophy2"
+
+      trophies.earned_trophies(np_communication_id: np_comm_id, platform: 'PS5')
+
+      expect(request).to have_received(:get).with(path)
+    end
+
+    it 'derives service name for PS3 platform' do
+      path = "/trophy/v1/users/me/npCommunicationIds/#{np_comm_id}/trophyGroups/all/trophies?npServiceName=trophy"
+
+      trophies.earned_trophies(np_communication_id: np_comm_id, platform: 'PS3')
+
+      expect(request).to have_received(:get).with(path)
+    end
+
+    it 'derives service name for PSVita platform' do
+      path = "/trophy/v1/users/me/npCommunicationIds/#{np_comm_id}/trophyGroups/all/trophies?npServiceName=trophy"
+
+      trophies.earned_trophies(np_communication_id: np_comm_id, platform: 'PSVita')
+
+      expect(request).to have_received(:get).with(path)
+    end
+
+    it 'derives service name for PC platform' do
+      path = "/trophy/v1/users/me/npCommunicationIds/#{np_comm_id}/trophyGroups/all/trophies?npServiceName=trophy2"
+
+      trophies.earned_trophies(np_communication_id: np_comm_id, platform: 'PC')
+
+      expect(request).to have_received(:get).with(path)
+    end
+
+    it 'accepts a custom trophy_group_id' do
+      path = "/trophy/v1/users/me/npCommunicationIds/#{np_comm_id}/trophyGroups/001/trophies?"
+
+      trophies.earned_trophies(np_communication_id: np_comm_id, trophy_group_id: '001')
+
+      expect(request).to have_received(:get).with(path)
+    end
+
+    it 'raises an error if both np_service_name and platform are provided' do
+      expect do
+        trophies.earned_trophies(np_communication_id: np_comm_id, platform: 'PS4', np_service_name: 'trophy')
+      end.to raise_error(ArgumentError, /Provide either np_service_name or platform, not both/)
+    end
+
+    it 'includes limit and offset when provided' do
+      path = "/trophy/v1/users/me/npCommunicationIds/#{np_comm_id}/trophyGroups/all/trophies?limit=10&offset=5"
+
+      trophies.earned_trophies(np_communication_id: np_comm_id, limit: 10, offset: 5)
 
       expect(request).to have_received(:get).with(path)
     end