+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Added `earned_trophies` method to retrieve trophy earned status for a user on a specific title
+
 ## [0.1.0] - 2025-11-01
 
 - Initial release